Move foreground call to inside NotifyActionInvokedAsync
