+++ tools/Dashboard/DevHome.Dashboard/ComSafeWidgetObjects/ComSafeWidget.cs
@@ -11,6 +11,7 @@ using Microsoft.Windows.Widgets;
 using Microsoft.Windows.Widgets.Hosts;
 using Serilog;
 using Windows.Foundation;
+using WinRT;
 
 namespace DevHome.Dashboard.ComSafeWidgetObjects;
 
@@ -186,6 +187,16 @@ public class ComSafeWidget
             try
             {
                 await LazilyLoadOopWidget();
+                try
+                {
+                    CoAllowSetForegroundWindow(_oopWidget);
+                }
+                catch (Exception ex)
+                {
+                    // If CoAllowSetForegroundWindow fails, we should still continue with the call to NotifyActionInvokedAsync.
+                    _log.Warning(ex, $"Call to CoAllowSetForegroundWindow failed");
+                }
+
                 await Task.Run(async () => await _oopWidget.NotifyActionInvokedAsync(verb, data));
                 return;
             }
@@ -363,5 +374,16 @@ public class ComSafeWidget
     public Widget GetUnsafeWidgetObject()
     {
         return _oopWidget;
+    }
+
+    // CoAllowSetForegroundWindow must be called on a raw COM interface, not a .NET CCW, in order to work correctly, since
+    // the underlying functionality is implemented by COM runtime and the object itself. CoAllowSetForegroundWindow wrapper
+    // below takes a WinRT object and extracts the raw COM interface pointer from it before calling native CoAllowSetForegroundWindow.
+    [DllImport("ole32.dll", ExactSpelling = true, PreserveSig = false)]
+    private static extern void CoAllowSetForegroundWindow(IntPtr pUnk, IntPtr lpvReserved);
+
+    private void CoAllowSetForegroundWindow(Widget widget)
+    {
+        CoAllowSetForegroundWindow(((IWinRTObject)widget).NativeObject.ThisPtr, 0);
     }
 }
